refactor(puzzles-data): split data access and transforms
